+++ plugin/src/net/groboclown/idea/p4ic/v2/server/cache/P4ChangeListValue.java
@@ -84,4 +84,9 @@ public class P4ChangeListValue {
         }
         return id;
     }
+
+    @Override
+    public String toString() {
+        return clientServerId + "@" + getChangeListId();
+    }
 }